+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## UNRELEASED
 
 ### Added
+
+- Added support for `quint test --backend=rust` to run tests using the Rust backend (#1862)
+
 ### Changed
 
 - Moved LSP-server compilation and analysis to its own worker thread (1876)